+++ src/Resources/contao/languages/de/tl_settings.php
@@ -1,0 +1,22 @@
+<?php if (!defined('TL_ROOT')) die('You cannot access this file directly!');
+
+/**
+ * Contao Open Source CMS
+ *
+ * Copyright (C) 2005-2013 Leo Feyer
+ *
+ * @package   fen
+ * @author    Frank Hoppe
+ * @license   GNU/LGPL
+ * @copyright Frank Hoppe 2013
+ */
+
+/**
+ * legends
+ */
+$GLOBALS['TL_LANG']['tl_settings']['spielerregister_legend']                      = 'Spielerregister';
+
+/**
+ * fields
+ */
+$GLOBALS['TL_LANG']['tl_settings']['spielerregister_imageSize'] = array('Bildgröße', 'Größe der Vorschaubilder');